Discover stories with Jsoup
Using Jsoup library, we look at author pages and story pages on fanfiction.net
From author pages, we find their stories and their favorite stories.
From story pages, we find how many more chapters there are for the story.

Ideally, we should not load the story page, but instead the author page when trying
to get information about a story, so that we minimize screwing up the view counts for authors.

Currently we will print out the discovered stories and chapters as a list.